+++ sdk/keyvault/Azure.Security.KeyVault.Certificates/tests/CertificateClientLiveTests.cs
@@ -114,6 +114,9 @@ namespace Azure.Security.KeyVault.Certificates.Tests
 
             RegisterForCleanup(certName);
 
+            // Give the service time to process the start request.
+            await DelayAsync(TimeSpan.FromSeconds(2));
+
             OperationCanceledException ex = null;
             try
             {
@@ -151,6 +154,9 @@ namespace Azure.Security.KeyVault.Certificates.Tests
             CertificateOperation operation = await Client.StartCreateCertificateAsync(certName, certificatePolicy);
 
             RegisterForCleanup(certName);
+
+            // Give the service time to process the start request.
+            await DelayAsync(TimeSpan.FromSeconds(2));
 
             OperationCanceledException ex = null;
             try
@@ -262,11 +268,9 @@ namespace Azure.Security.KeyVault.Certificates.Tests
                 RegisterForCleanup(certName);
 
                 using CancellationTokenSource cts = new CancellationTokenSource(DefaultCertificateOperationTimeout);
-                TimeSpan pollingInterval = Mode == RecordedTestMode.Playback ? TimeSpan.Zero : KeyVaultTestEnvironment.DefaultPollingInterval;
-
                 while (!operation.HasCompleted)
                 {
-                    await Task.Delay(pollingInterval, cts.Token);
+                    await Task.Delay(PollingInterval, cts.Token);
                     await operation.UpdateStatusAsync(cts.Token);
                 }
 
@@ -355,9 +359,7 @@ namespace Azure.Security.KeyVault.Certificates.Tests
 
             // Need to call the real async wait method or the sync version of this test fails because it's using the instrumented Client directly.
             using CancellationTokenSource cts = new CancellationTokenSource(DefaultCertificateOperationTimeout);
-            TimeSpan pollingInterval = Mode == RecordedTestMode.Playback ? TimeSpan.Zero : KeyVaultTestEnvironment.DefaultPollingInterval;
-
-            await operation.WaitForCompletionAsync(pollingInterval, cts.Token);
+            await operation.WaitForCompletionAsync(PollingInterval, cts.Token);
 
             Assert.IsTrue(operation.HasCompleted);
             Assert.IsTrue(operation.HasValue);