+++ sdk/monitor/Azure.Monitor.OpenTelemetry.Exporter/src/TelemetryPartB.cs
@@ -45,7 +45,7 @@ namespace Azure.Monitor.OpenTelemetry.Exporter
             }
 
             var statusCode = AzMonList.GetTagValue(ref monitorTags.PartBTags, SemanticConventions.AttributeHttpStatusCode)?.ToString() ?? "0";
-            var success = activity.GetStatus() != Status.Error;
+            var success = activity.GetStatus().StatusCode != StatusCode.Error;
             var request = new RequestData(2, activity.Context.SpanId.ToHexString(), activity.Duration.ToString("c", CultureInfo.InvariantCulture), success, statusCode)
             {
                 Name = activity.DisplayName,
@@ -64,7 +64,7 @@ namespace Azure.Monitor.OpenTelemetry.Exporter
             var dependency = new RemoteDependencyData(2, activity.DisplayName, activity.Duration.ToString("c", CultureInfo.InvariantCulture))
             {
                 Id = activity.Context.SpanId.ToHexString(),
-                Success = activity.GetStatus() != Status.Error
+                Success = activity.GetStatus().StatusCode != StatusCode.Error
             };
 
             switch (monitorTags.activityType)