+++ sdk/network/Azure.ResourceManager.Network/tests/Tests/VirtualNetworkPeeringTests.cs
@@ -11,7 +11,7 @@ using Azure.ResourceManager.Network.Tests.Helpers;
 using NUnit.Framework;
 using SubResource = Azure.ResourceManager.Network.Models.SubResource;
 
-namespace Azure.ResourceManager.Network.Tests.Tests
+namespace Azure.ResourceManager.Network.Tests
 {
     public class VirtualNetworkPeeringTests : NetworkServiceClientTestBase
     {