+++ sdk/tables/Azure.Data.Tables/tests/samples/Sample2_CreateDeleteEntities.cs
@@ -8,7 +8,6 @@ using Azure.Data.Tables.Tests;
 
 namespace Azure.Data.Tables.Samples
 {
-    [LiveOnly]
     public partial class TablesSamples : TablesTestEnvironment
     {
         [Test]
@@ -17,7 +16,7 @@ namespace Azure.Data.Tables.Samples
             string storageUri = StorageUri;
             string accountName = StorageAccountName;
             string storageAccountKey = PrimaryStorageAccountKey;
-            string tableName = "OfficeSupplies2p1";
+            string tableName = "OfficeSupplies2p1" + _random.Next();
             string partitionKey = "Stationery";
             string rowKey = "A1";
             string rowKeyStrong = "B1";